\name{rmethrafo-package}
\alias{rmethrafo-package}
\alias{rmethrafo}
\docType{package}
\title{methrafo}
\description{MeDIP-Seq methylation estimation}
\details{
  \tabular{ll}{
    Package: \tab rmethrafo\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2017-06-20\cr
    License: \tab MIT\cr
  }
}
\author{
Ziv Bar-Joseph <zivbj@cs.cmu.edu>
Jun Ding <jund@cs.cmu.edu>
}
